BACKEND/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
bs4==0.0.2
build==1.2.1
cachetools==5.4.0
cattrs==23.2.3
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
cron-descriptor==1.4.3
cryptography==43.0.0
cssselect==1.2.0
dataclasses-json==0.6.7
decouple==0.0.7
Deprecated==1.2.14
distro==1.9.0
Django==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-filter==24.3
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.111.1
fastapi-cli==0.0.4
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.2
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwt==1.3.1
kombu==5.3.7
kubernetes==30.1.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.24
langchain-google-genai==1.0.8
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.93
lxml==5.2.2
lxml_html_clean==0.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==0.28.0
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
posthog==3.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyMySQL==1.1.1
pyparsing==3.1.2
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pytest==8.3.1
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.4
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-cache==1.2.1
requests-file==2.1.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
starlette==0.37.2
streamlit==1.37.0
sympy==1.13.1
tenacity==8.5.0
tiktoken==0.7.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
